Quick fix scoring 'dont know' selections as 0
This is a quick fix, since if the user selects 'Don't know/Doesn't
apply' as an answer it means the average score for this dimension
isn't calculated.

This happens because you can't average a mixture of strings and
numbers.

To fix this, I've simply suggested 'Don't know's' now are recorded
as zero.